fix money format on customers

+++ database/migrations/2023_02_19_134001_create_customers_table.php
@@ -13,8 +13,8 @@ return new class () extends Migration
         Schema::create('customers', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('balance')->default(0);
-            $table->integer('paid_to_date')->default(0);
+            $table->decimal('balance')->default(0);
+            $table->decimal('paid_to_date')->default(0);
             $table->timestamps();
         });
     }
